+++ jabgui/src/main/java/org/jabref/gui/libraryproperties/saving/SavingPropertiesViewModel.java
@@ -7,14 +7,17 @@ import java.util.Set;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ListProperty;
+import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleListProperty;
+import javafx.beans.property.SimpleObjectProperty;
 import javafx.collections.FXCollections;
 
 import org.jabref.gui.commonfxcontrols.SortCriterionViewModel;
 import org.jabref.gui.libraryproperties.PropertiesTabViewModel;
 import org.jabref.logic.cleanup.FieldFormatterCleanup;
 import org.jabref.logic.cleanup.FieldFormatterCleanupActions;
+import org.jabref.logic.journals.AbbreviationType;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.field.Field;
 import org.jabref.model.entry.field.FieldFactory;
@@ -46,6 +49,9 @@ public class SavingPropertiesViewModel implements PropertiesTabViewModel {
     // FieldFormatterCleanupsPanel
     private final BooleanProperty cleanupsDisableProperty = new SimpleBooleanProperty();
     private final ListProperty<FieldFormatterCleanup> cleanupsProperty = new SimpleListProperty<>(FXCollections.emptyObservableList());
+
+    // Journal abbreviation on save
+    private final ObjectProperty<AbbreviationType> journalAbbreviationOnSaveProperty = new SimpleObjectProperty<>();
 
     private final BibDatabaseContext databaseContext;
     private final MetaData initialMetaData;
@@ -98,6 +104,8 @@ public class SavingPropertiesViewModel implements PropertiesTabViewModel {
             cleanupsDisableProperty.setValue(!defaultSaveActions.isEnabled());
             cleanupsProperty.setValue(FXCollections.observableArrayList(defaultSaveActions.getConfiguredActions()));
         });
+
+        journalAbbreviationOnSaveProperty.setValue(initialMetaData.getLibraryAbbreviationType().orElse(null));
     }
 
     @Override
@@ -135,6 +143,13 @@ public class SavingPropertiesViewModel implements PropertiesTabViewModel {
             } else {
                 newMetaData.setSaveOrder(newSaveOrder);
             }
+        }
+
+        AbbreviationType abbreviationType = journalAbbreviationOnSaveProperty.getValue();
+        if (abbreviationType != null) {
+            newMetaData.setLibraryAbbreviationType(abbreviationType);
+        } else {
+            newMetaData.clearLibraryAbbreviationType();
         }
 
         databaseContext.setMetaData(newMetaData);
@@ -178,5 +193,9 @@ public class SavingPropertiesViewModel implements PropertiesTabViewModel {
 
     public ListProperty<FieldFormatterCleanup> cleanupsProperty() {
         return cleanupsProperty;
+    }
+
+    public ObjectProperty<AbbreviationType> journalAbbreviationOnSaveProperty() {
+        return journalAbbreviationOnSaveProperty;
     }
 }